requirements.txt
vosk~=0.3.45
soundfile~=0.12.1
pydub~=0.25.1
python-dotenv~=1.0.1
numpy~=1.26.4
tensorflow~=2.16.1
googletrans~=4.0.0rc1
transformers==4.16.2